*new function* clearUserQuizData() - this is called when a user jumps off the chair while the chair is in the quizzing state.  It will reset the appropriate vars so that a new user can take the quiz.  It will also re-randomize the question ids.
This function gets called in two places.  One, in the check_quiz state, where
we firstly initialize the variables, and also in the changed event in the quizzing state
if a user jumps off a chair.  In the changed event handler in the quizzing state,
I also tell the quiz to go to the ready state if a user jumps off (after first calling this function)

*Various Linked messages*
I also added a new linked message on every state change.  These will be useful
for other effects scripts in the future.  Ie: When we are in the quizzing state
trigger a "3 -2-1 GO!" sound etc.

*File Name Change*
renamed to sloodle_mod_quiz-2.0.lslp to reflect our version change. Also added the lslp
extension because the Eclipse plugin LSL PLUS doesnt work with the .lsl extension
even if that extension is added to the eclipse editor.  The extension change
shouldmake a difference to hard core vi users, however perhaps the file name change will.  Need to consult
Edmund on the name change
--added sloodle_status_codes
This files lists all the status codes we use for sloodle.
They have been written in LSL format so that you can plunk them into your source
code if needed.