Quizzer was not advancing to the next question, nor storing scores
Fixed issue where questions were not advancing.  Quizzer now works for
multiple users! TODO, add an option so that the quizzer does not ask
questions continuously, but instead allows the admin to make a setting
so that the quizzer only asks a question when touched.  Also, add
appropriate llMessageLinked to other scripts can hook into the events of
this script