Fixed bug where an extra dialog was showing after each question
the notify_server function (called in LSL when a user responds to a
question) was not submitting a user key.  Fixed this.  Also fixed all
llhttprequests to utlize the global sloodlehttpvars parameter, instead
of redfining it each tim. Replaced an un-translated llStetText with
sloodle_translate_request function instead - now it can be translated!
Added a request_descriptor in each llhttprequest, which is now
sent/returned in the sloodle_status_fields, so when the lsl script
receives a response from the server, it will know which lsl task
initiated the request.