modified multi_user_quiz to await messages from the platform script to load a quiz
once the rezzer receives a regionsay from a platform, it will pass it on
to the multi_user quiz via the
SLOODLE_CHANNEL_QUIZ_STATE_ENTRY_LOAD_QUIZ_FOR_USER channel, the
multi_user_quiz should then try to load the quiz for the user.  Also
added a hover text prim for the main parent hex, to display status text
to the users while it is waiting for all of the other scripts to finish
loading the initial quiz data.  TODO: react to multi_user_quiz script.