Updated animations for pie_slices, and orbs
removed p=0 from pie_slices, this was actually an orb animation.
Updated all orb animations so that they dimple, and hollow out so that
they are completely hidden when pie_slices are open and closed.  Also
updated their animations so that they are in correct orientation when
rezzed from the rezzer (rotated root prim to 0,0,0)