hex now opens th incorrect pie_slices, added hovertext to pie_slices with avatars standing over when timer is running
removed getmy_coord - not needed anymore because we changed the root
prim to the center orb (previously orb 0).

created new function which sets the message of all pie_slices
created the feature that when the timer is counting down, a sensor
starts an lists all avatars names overtop the pie_slice they are
standing on.

When time is up, all incorrect pie_slices open